services/biobert/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.7.0
python-dotenv==1.0.0
# BioBERT NER via HuggingFace
transformers==4.52.4
torch==2.7.1
# Fast tokenization
tokenizers>=0.19.0
# Numpy (required by transformers)
numpy>=1.24.0
